pyproject.toml
[tool.poetry]
name = "my_analysis"
version = "0.1.0"
description = ""
authors = ["Maxim Laurijssen <maxim.laurijssen@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.9"
python-box = "^5.3.0"
numpy = "^1.21.1"
pandas = "^1.3.1"
openpyxl = "^3.0.7"
PyYAML = "^5.4.1"
scikit-learn = "^0.24.2"
semopy = "^2.3.5"
pandarallel = "^1.5.2"
tqdm = "^4.62.0"
statsmodels = "^0.13.1"
scipy = "^1.7.3"
weasyprint = "^53.4"
pyreadstat = "^1.1.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
